requirements.txt
streamlit
google-generativeai
transformers
torch
torchvision
pillow
fpdf
openai
requests
opencv-python